rename: sidebar sideNavbar로 네이밍 변경
