+++ src/components/layout/sideNavbar/PageListToggleButton.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import { MenuItem, MenuItemProps, styled } from "@mui/material";
 import ToggleIcon from "@/assets/toggle.svg?react";
-import NavigateButton from "@/components/layout/sidebar/NavigateButton";
+import NavigateButton from "@/components/layout/sideNavbar/NavigateButton";
 
 export type NavigatePage = { label: string; path: string };
